# Conflicts:
#	SofaKernel/framework/sofa/helper/CMakeLists.txt
#	SofaKernel/framework/sofa/helper/gl/VideoRecorder.cpp

+++ SofaKernel/framework/sofa/core/objectmodel/AspectPool.h
@@ -1,6 +1,6 @@
 /******************************************************************************
 *       SOFA, Simulation Open-Framework Architecture, development version     *
-*                (c) 2006-2018 INRIA, USTL, UJF, CNRS, MGH                    *
+*                (c) 2006-2019 INRIA, USTL, UJF, CNRS, MGH                    *
 *                                                                             *
 * This program is free software; you can redistribute it and/or modify it     *
 * under the terms of the GNU Lesser General Public License as published by    *
@@ -104,8 +104,8 @@ public:
 
     AspectRef getAspect(int id);
 
-    int nbAspects() const { return (int) aspects.size(); }
-    int getAspectCounter(int id) const { return aspects[id]->counter; }
+    int nbAspects() const { return int(aspects.size()); }
+    int getAspectCounter(int id) const { return aspects[size_t(id)]->counter; }
 
     friend class Aspect;
